requirements.txt
anyio==3.7.1
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
PyJWT==2.8.0
pymongo==4.6.3
python-dotenv==1.0.1
sniffio==1.3.1
tomli==2.0.1
Werkzeug==3.0.2
zope.event==5.0
zope.interface==6.3
